Altered the beacon one red program
I altered the beacon one red program to fix it's orientation and be more
aligned with the goal before it launches the particles. Then it will
strafe left until the ramp, and drive forward until it sees the white
line of the first beacon. It currently needs to mave backwards a small
amount before strafing left, and the line sensing needs to be improved.